# docs/swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Claude Code container orchestration API - secure, isolated AI execution",
        "title": "Stromboli API",
        "termsOfService": "https://github.com/tomblanc/stromboli",
        "contact": {
            "name": "API Support",
            "url": "https://stromboli/issues"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/jobs/{id}": {
            "get": {
                "description": "Returns the status and result of an async job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Get job status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api.JobResponse"
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Cancels a pending or running job",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "jobs"
                ],
                "summary": "Cancel job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Job ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Job not found",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    },
                    "409": {
                        "description": "Job cannot be cancelled",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    }
                }
            }
        },
...
        "/run/async": {
            "post": {
                "description": "Starts Claude Code execution asynchronously and returns a job ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "execution"
                ],
                "summary": "Run Claude async",
                "parameters": [
                    {
                        "description": "Run request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunRequest"
                        }
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "$ref": "#/definitions/internal_api.AsyncRunResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    },
                    "503": {
                        "description": "Claude not configured",
                        "schema": {
                            "$ref": "#/definitions/internal_api.RunResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "internal_api.AsyncRunResponse": {
            "description": "Response from starting an async Claude execution",
            "type": "object",
            "properties": {
                "job_id": {
                    "type": "string",
                    "example": "job-abc123def456"
                },
                "session_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "internal_api.JobResponse": {
            "description": "Job status and result",
            "type": "object",
            "properties": {
                "crash_info": {
                    "$ref": "#/definitions/stromboli_internal_job.CrashInfo"
                },
                "created_at": {
                    "type": "string",
                    "example": "2024-01-15T10:30:00Z"
                },
                "error": {
                    "type": "string"
                },
                "id": {
                    "type": "string",
                    "example": "job-abc123def456"
                },
                "output": {
                    "type": "string",
                    "example": "Hello!"
                },
                "session_id": {
                    "type": "string",
                    "example": "sess-abc123def456"
                },
                "status": {
                    "allOf": [
                        {
                            "$ref": "#/definitions/stromboli_internal_job.Status"
                        }
                    ],
                    "example": "running"
                },
                "structured_output": {
                    "type": "object"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2024-01-15T10:31:00Z"
                }
            }
        },
...
        "internal_api.RunRequest": {
            "description": "Request to execute Claude Code in an isolated container",
            "type": "object",
            "required": [
                "prompt"
            ],
            "properties": {
                "claude": {
                    "description": "Claude configuration - all CLI options exposed",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stromboli_internal_types.ClaudeOptions"
                        }
                    ]
                },
                "podman": {
                    "description": "Podman configuration",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stromboli_internal_types.PodmanOptions"
                        }
                    ]
                },
                "prompt": {
                    "description": "Required: the prompt to send to Claude",
                    "type": "string",
                    "example": "Analyze this code and suggest improvements"
                },
                "webhook_url": {
                    "description": "Webhook URL to notify when job completes (async only)",
                    "type": "string",
                    "example": "https://example.com/webhook"
                },
                "workdir": {
                    "description": "Working directory inside the container where Claude will spawn\nUse podman.volumes to mount host paths into the container",
                    "type": "string",
                    "example": "/workspace"
                }
            }
        },
        "internal_api.RunResponse": {
            "description": "Response from Claude execution",
            "type": "object",
            "properties": {
                "crash_info": {
                    "description": "Crash details (when status is \"crashed\")",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stromboli_internal_job.CrashInfo"
                        }
                    ]
                },
                "error": {
                    "description": "Error message (when failed)",
                    "type": "string",
                    "example": ""
                },
                "id": {
                    "description": "Unique run identifier",
                    "type": "string",
                    "example": "run-abc123def456"
                },
                "output": {
                    "description": "Claude's raw output (full CLI output, including JSON envelope when output_format=json)",
                    "type": "string",
                    "example": "Here is my analysis..."
                },
                "session_id": {
                    "description": "Session ID for conversation continuation",
                    "type": "string",
                    "example": "sess-abc123def456"
                },
                "status": {
                    "description": "Execution status: completed, crashed, error",
                    "type": "string",
                    "example": "completed"
                },
                "structured_output": {
                    "description": "Extracted structured_output from Claude's JSON envelope (when json_schema is used).\nConvenience field so callers don't need to parse the envelope themselves.",
                    "type": "object"
                },
                "usage": {
                    "description": "Token usage totals + estimated USD cost. Populated best-effort from the\nsession's JSONL; nil when the session file isn't readable yet or no\nassistant message reported usage.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stromboli_internal_job.Usage"
                        }
                    ]
                }
            }
        },
...
        "stromboli_internal_job.CrashInfo": {
            "type": "object",
            "properties": {
                "exit_code": {
                    "description": "Exit code (if available)",
                    "type": "integer"
                },
                "partial_output": {
                    "description": "Partial output captured before crash",
                    "type": "string"
                },
                "reason": {
                    "description": "Human-readable crash reason",
                    "type": "string"
                },
                "signal": {
                    "description": "Signal that killed the process (SIGSEGV, SIGKILL, etc.)",
                    "type": "string"
                },
                "task_completed": {
                    "description": "Whether the task appeared to complete before crashing",
                    "type": "boolean"
                }
            }
        },
        "stromboli_internal_job.Status": {
            "type": "string",
            "enum": [
                "pending",
                "running",
                "completed",
                "failed",
                "crashed",
                "cancelled"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusRunning",
                "StatusCompleted",
                "StatusFailed",
                "StatusCrashed",
                "StatusCancelled"
            ]
        },
        "stromboli_internal_job.Usage": {
            "type": "object",
            "properties": {
                "cache_creation_input_tokens": {
                    "type": "integer"
                },
                "cache_read_input_tokens": {
                    "type": "integer"
                },
                "estimated_cost_usd": {
                    "description": "EstimatedCostUSD is best-effort: tokens × Anthropic's published\nper-1M rates, summed across all four buckets. Zero when the model\nisn't recognized or no usage was reported.",
                    "type": "number"
                },
                "input_tokens": {
                    "type": "integer"
                },
                "model": {
                    "description": "Model identifier reported by Claude (e.g. \"claude-opus-4-7-20260101\").\nEmpty when the model is unknown — cost estimation skipped in that case.",
                    "type": "string"
                },
                "output_tokens": {
                    "type": "integer"
                },
                "total_tokens": {
                    "type": "integer"
                }
            }
        },
        "stromboli_internal_types.ClaudeOptions": {
            "description": "All available Claude CLI options for headless execution",
            "type": "object",
            "properties": {
                "add_dirs": {
                    "description": "Additional directories for tool access",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "agent": {
                    "description": "Agent for current session",
                    "type": "string",
                    "example": "reviewer"
                },
                "agents": {
                    "description": "Custom agents definition (JSON object)",
                    "type": "object",
                    "additionalProperties": {}
                },
                "allow_dangerously_skip_permissions": {
                    "description": "Enable bypass as an option without enabling by default",
                    "type": "boolean",
                    "example": false
                },
                "allowed_tools": {
                    "description": "Allowed tools with patterns (e.g., \"Bash(git:*)\")",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Bash(git:*)",
                        "Read"
                    ]
                },
                "append_system_prompt": {
                    "description": "Append to default system prompt",
                    "type": "string",
                    "example": "Focus on security best practices"
                },
                "betas": {
                    "description": "Beta headers for API requests",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "continue": {
                    "description": "Continue most recent conversation in workspace (ignores session_id)",
                    "type": "boolean",
                    "example": false
                },
                "dangerously_skip_permissions": {
                    "description": "Bypass all permission checks (use in sandboxed environments only)",
                    "type": "boolean",
                    "example": true
                },
                "debug": {
                    "description": "Debug mode with optional category filter",
                    "type": "string",
                    "example": "api,hooks"
                },
                "disable_slash_commands": {
                    "description": "Disable all slash commands/skills",
                    "type": "boolean",
                    "example": false
                },
                "disallowed_tools": {
                    "description": "Denied tools",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Write"
                    ]
                },
                "fallback_model": {
                    "description": "Fallback model when default is overloaded",
                    "type": "string",
                    "example": "haiku"
                },
                "files": {
                    "description": "File resources (format: file_id:path)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "fork_session": {
                    "description": "Create new session ID when resuming",
                    "type": "boolean",
                    "example": false
                },
                "include_partial_messages": {
                    "description": "Include partial message chunks (stream-json only)",
                    "type": "boolean",
                    "example": false
                },
                "input_format": {
                    "description": "Input format: text, stream-json",
                    "type": "string",
                    "example": "text"
                },
                "json_schema": {
                    "description": "JSON Schema for structured output validation",
                    "type": "string",
                    "example": "{\"type\":\"object\"}"
                },
                "max_budget_usd": {
                    "description": "Maximum dollar amount for API calls (pointer to distinguish 0 from unset)",
                    "type": "number",
                    "example": 5
                },
                "max_turns": {
                    "description": "Maximum number of agentic turns (0 = unlimited, nil = use CLI default)",
                    "type": "integer",
                    "example": 30
                },
                "mcp_configs": {
                    "description": "MCP server config files or JSON strings",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "model": {
                    "description": "Model alias (sonnet, opus, haiku) or full name",
                    "type": "string",
                    "example": "sonnet"
                },
                "no_persistence": {
                    "description": "Don't save session to disk",
                    "type": "boolean",
                    "example": false
                },
                "output_format": {
                    "description": "Output format: text, json, stream-json",
                    "type": "string",
                    "example": "json"
                },
                "permission_mode": {
                    "description": "Permission mode: acceptEdits, bypassPermissions, default, delegate, dontAsk, plan",
                    "type": "string",
                    "example": "bypassPermissions"
                },
                "plugin_dirs": {
                    "description": "Plugin directories",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "replay_user_messages": {
                    "description": "Re-emit user messages on stdout",
                    "type": "boolean",
                    "example": false
                },
                "resume": {
                    "description": "Resume an existing session (requires session_id)",
                    "type": "boolean",
                    "example": true
                },
                "session_id": {
                    "description": "Session ID (UUID) - used for both new and resumed sessions",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "setting_sources": {
                    "description": "Setting sources to load: user, project, local",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "user",
                        "project"
                    ]
                },
                "settings": {
                    "description": "Path to settings JSON file or JSON string",
                    "type": "string"
                },
                "strict_mcp_config": {
                    "description": "Only use MCP servers from mcp_configs",
                    "type": "boolean",
                    "example": false
                },
                "system_prompt": {
                    "description": "Replace default system prompt",
                    "type": "string",
                    "example": "You are a senior Go developer"
                },
                "tools": {
                    "description": "Built-in tools (\"\", \"default\", or specific names)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "Bash",
                        "Read",
                        "Edit"
                    ]
                },
                "verbose": {
                    "description": "Enable verbose mode",
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "stromboli_internal_types.EnvironmentConfig": {
            "description": "Runtime environment configuration (single container or compose)",
            "type": "object",
            "properties": {
                "build_timeout": {
                    "description": "Optional build timeout override for compose (e.g., \"15m\")\nIf not specified, uses server default (10m)",
                    "type": "string",
                    "example": "15m"
                },
                "path": {
                    "description": "Path to compose file (required when type=\"compose\")\nMust be an absolute path ending in .yml or .yaml",
                    "type": "string",
                    "example": "/home/user/project/docker-compose.yml"
                },
                "service": {
                    "description": "Service name where Claude will run (required when type=\"compose\")",
                    "type": "string",
                    "example": "dev"
                },
                "type": {
                    "description": "Type of environment: \"\" (default single container) or \"compose\"",
                    "type": "string",
                    "example": "compose"
                }
            }
        },
...
        "stromboli_internal_types.PodmanOptions": {
            "description": "Podman container mount configuration",
            "type": "object",
            "properties": {
                "cpu_shares": {
                    "description": "CPU shares (relative weight, default 1024; pointer to distinguish 0 from unset)",
                    "type": "integer",
                    "example": 512
                },
                "cpus": {
                    "description": "CPU limit (e.g., \"0.5\", \"2\")",
                    "type": "string",
                    "example": "1"
                },
                "environment": {
                    "description": "Environment specifies a compose-based multi-service environment.\nWhen set, the agent runs inside the specified service of the compose stack\ninstead of a standalone container.",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stromboli_internal_types.EnvironmentConfig"
                        }
                    ]
                },
                "image": {
                    "description": "Container image override (must match allowed patterns)",
                    "type": "string",
                    "example": "python:3.12"
                },
                "lifecycle": {
                    "description": "Lifecycle hooks for running commands at specific container events",
                    "allOf": [
                        {
                            "$ref": "#/definitions/stromboli_internal_types.LifecycleHooks"
                        }
                    ]
                },
                "memory": {
                    "description": "Memory limit (e.g., \"512m\", \"1g\")",
                    "type": "string",
                    "example": "512m"
                },
                "secrets_env": {
                    "description": "Secrets to inject as environment variables (env_var_name -\u003e podman_secret_name)\nThe Podman secret must exist beforehand (created via `podman secret create`)\nExample: {\"GH_TOKEN\": \"github-token\"} mounts secret \"github-token\" as env var GH_TOKEN",
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "timeout": {
                    "description": "Container timeout (e.g., \"5m\", \"1h\", \"30s\")",
                    "type": "string",
                    "example": "5m"
                },
                "volumes": {
                    "description": "Volume mounts (host:container or host:container:options format)",
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "/data:/data:ro"
                    ]
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "OAuth2 Bearer token (future)",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}